OpenWeatherMap now requires to use the API Key with every API call
from any type of account, including FREE subscription. Currently, the
only way to deal with this is to explicitly pass the API key with every
invocation of any method of the OpenWeatherMap class. This modification
allows to pass the API key to the class contractor and save it as a
private member, thus eliminating the need to pass the key to the
individual calls. However, for backward compatibility, the API key can
be passed to the individual method calls which effectively overrides the
API key that persists in the class instance.